remove test_fb_* test modules, keep tests that are still useful.
